requirements.txt
julian==0.14
PyAstronomy==0.20.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
